parser_app/pyproject.toml
[project]
name = "sconewsbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
  "boilerpy3>=1.0.7",
  "httpx>=0.28.1",
  "newspaper4k>=0.9.3.1",
  "lxml[html_clean]",
  "typing-extensions",
  "pymorphy3",
  "pymorphy3-dicts-ru",
  "requests",
  "setuptools",
  "wheel",
  "python-dotenv>=1.0.1",
  "openai>=1.59.9",
  "asyncpg>=0.30.0",
  "pydantic>=2.10.5",
  "tortoise-orm>=0.24.0",
]

[tool.ruff]
exclude = ["**/.venv/**", "**/migrations/**", "**/sandbox/**"]
lint.select = ["ALL"]
lint.ignore = ["D203", "D213", "TC001", "ARG001", "S101", "TC002", "D100", "E501"]
line-length = 79

[tool.ruff.lint.mccabe]
max-complexity = 4

[tool.black]
line-length = 79